捡道具，修人物动画机，修ui，实现真cd，修配置表
Former-commit-id: c38b999294f8eafc7e7bf4236b84b80ac3ace7c9

+++ Assets/Scripts/PropConfig.cs
@@ -21,7 +21,7 @@ public class PropData
     [Header("改变子弹速度")]
     public float changeBulletSpeed;
     [Header("改变子弹伤害")]
-    public int changeDamage;
+    public float changeDamage;
     [Header("改变移动速度")]
     public float changeSpeed;
     [Header("改变子弹效果")]
